dj_admin/requirements.txt
asgiref==3.7.2
cffi==1.16.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
Django==4.2.4
django-crispy-forms==2.0
django-password-validators==1.7.1
django-session-timeout==0.1.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
